db: moved db services into its own dir
